Minor fix to histogram binning
I saw this minor error and figured I'd figure out how these pull requests work. 
It's clear from the original line that you meant to have the histogram use less than 100 bins if there is not enough data to justify it, however the parenthesis is set slightly wrong for this to work correctly. ```len(a/10)``` is always the same as ```len(a)```, but different from ```len(a)/10```. Oh, and the division by 10 must be an integer division, as ```hist``` barfs if you ask for a decimal number of bins.